refactored ssl loading into ravenextensions.cs
